requirements.txt
# Core dependencies for Forge Kontext Assistant V2

# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0

# Transformers for Florence-2
transformers>=4.35.0

# Image processing
Pillow>=9.0.0

# Optional but recommended
accelerate>=0.20.0
gradio>=3.50.0

# For development/testing
pytest>=7.0.0
requests>=2.28.0